+++ examples/cra/src/components/AddButton.js
@@ -1,6 +1,6 @@
-import React from 'react'
-import PropTypes from 'prop-types'
 import { useDataMutation } from '@dhis2/app-runtime'
+import PropTypes from 'prop-types'
+import React from 'react'
 
 const mutation = {
     resource: 'indicators',