+++ services/data/src/links/RestAPILink/queryToRequestOptions/multipartFormDataMatchers.test.ts
@@ -3,6 +3,7 @@ import {
     isMessageConversationAttachment,
     isStaticContentUpload,
     isAppInstall,
+    isSvgConversion,
 } from './multipartFormDataMatchers'
 
 describe('isFileResourceUpload', () => {
@@ -64,7 +65,7 @@ describe('isStaticContentUpload', () => {
 })
 
 describe('isAppInstall', () => {
-    it('returns true for a POST to "fileResources"', () => {
+    it('returns true for a POST to "apps"', () => {
         expect(
             isAppInstall('create', {
                 resource: 'apps',
@@ -75,6 +76,30 @@ describe('isAppInstall', () => {
         expect(
             isAppInstall('create', {
                 resource: 'notApps',
+            })
+        ).toEqual(false)
+    })
+})
+
+describe('isSvgConversion', () => {
+    it('returns true for a POST to "svg.png"', () => {
+        expect(
+            isSvgConversion('create', {
+                resource: 'svg.png',
+            })
+        ).toEqual(true)
+    })
+    it('returns true for a POST to "svg.pdf"', () => {
+        expect(
+            isSvgConversion('create', {
+                resource: 'svg.pdf',
+            })
+        ).toEqual(true)
+    })
+    it('retuns false for a POST to a different resource', () => {
+        expect(
+            isSvgConversion('create', {
+                resource: 'notSvg',
             })
         ).toEqual(false)
     })